+++ src/gsuite/clients/GSuiteDomainClient.ts
@@ -1,9 +1,20 @@
-import GSuiteClient from './GSuiteClient';
-
 import { admin_directory_v1 } from 'googleapis';
 
-export class GSuiteDomainClient extends GSuiteClient {
-  async iterateDomains(
+import GSuiteAdminClient from './GSuiteAdminClient';
+import { CreateGSuiteClientParams } from './GSuiteClient';
+
+export class GSuiteDomainClient extends GSuiteAdminClient {
+  constructor(params: CreateGSuiteClientParams) {
+    super({
+      ...params,
+      requiredScopes: [
+        'https://www.googleapis.com/auth/admin.directory.domain.readonly',
+        ...(params.requiredScopes || []),
+      ],
+    });
+  }
+
+  public async iterateDomains(
     callback: (data: admin_directory_v1.Schema$Domains) => Promise<void>,
   ): Promise<void> {
     const client = await this.getAuthenticatedServiceClient();